requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sentence-transformers
streamlit==1.31.0
pymupdf==1.23.21
python-docx==1.1.0
pandas==2.2.0
plotly==5.18.0
psutil==5.9.8
spacy==3.7.2
scikit-learn==1.4.0
seaborn
matplotlib